requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
dnspython==2.8.0
idna==3.11
psutil==7.2.1
pycparser==2.23
python-dateutil==2.9.0.post0
python-whois==0.9.6
pystyle==2.9
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
tqdm==4.67.1
urllib3==2.6.3
